Support multiple consumer types with configurable processing
