+++ consumer/processing.go
@@ -16,12 +16,10 @@ package consumer
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"time"
 
 	"github.com/Shopify/sarama"
-	"github.com/google/uuid"
 	"github.com/rs/zerolog/log"
 
 	"github.com/RedHatInsights/insights-results-aggregator/metrics"
@@ -45,6 +43,15 @@ var (
 		reportAttributeFingerprints, reportAttributeInfo, reportAttributeReports, reportAttributeSystem,
 	}
 )
+
+// MessageProcessor offers the interface for processing a received message
+//lint:ignore U1000 Ignore the golint warning about an exported method returning an unexported type.
+type MessageProcessor interface {
+	DeserializeMessage(messageValue []byte) (incomingMessage, error)
+	ParseMessage(consumer *KafkaConsumer, msg *sarama.ConsumerMessage) (incomingMessage, error)
+	ProcessMessage(consumer *KafkaConsumer, msg *sarama.ConsumerMessage) (types.RequestID, incomingMessage, error)
+	ShouldProcess(consumer *KafkaConsumer, consumed *sarama.ConsumerMessage, parsed *incomingMessage) error
+}
 
 // Report represents report send in a message consumed from any broker
 type Report map[string]*json.RawMessage
@@ -131,7 +138,7 @@ var currentSchemaVersion = types.AllowedVersions{
 // 2:53PM DBG processing of message took '0.005895183' seconds offset=8 partition=0 topic=ccx.ocp.results
 // 2:53PM WRN request ID is missing, null or empty Operation=TrackPayload
 // 2:53PM INF Message consumed duration=6 offset=8
-func (consumer *OCPRulesConsumer) HandleMessage(msg *sarama.ConsumerMessage) error {
+func (consumer *KafkaConsumer) HandleMessage(msg *sarama.ConsumerMessage) error {
 	log.Info().
 		Int64(offsetKey, msg.Offset).
 		Int32(partitionKey, msg.Partition).
@@ -142,7 +149,7 @@ func (consumer *OCPRulesConsumer) HandleMessage(msg *sarama.ConsumerMessage) err
 	metrics.ConsumedMessages.Inc()
 
 	startTime := time.Now()
-	requestID, message, err := consumer.processMessage(msg)
+	requestID, message, err := consumer.MessageProcessor.ProcessMessage(consumer, msg)
 	timeAfterProcessingMessage := time.Now()
 	messageProcessingDuration := timeAfterProcessingMessage.Sub(startTime).Seconds()
 
@@ -183,7 +190,7 @@ func (consumer *OCPRulesConsumer) HandleMessage(msg *sarama.ConsumerMessage) err
 }
 
 // updatePayloadTracker
-func (consumer OCPRulesConsumer) updatePayloadTracker(
+func (consumer KafkaConsumer) updatePayloadTracker(
 	requestID types.RequestID,
 	timestamp time.Time,
 	orgID *types.OrgID,
@@ -199,7 +206,7 @@ func (consumer OCPRulesConsumer) updatePayloadTracker(
 }
 
 // sendDeadLetter - sends unprocessed message to dead letter queue
-func (consumer OCPRulesConsumer) sendDeadLetter(msg *sarama.ConsumerMessage) {
+func (consumer KafkaConsumer) sendDeadLetter(msg *sarama.ConsumerMessage) {
 	if consumer.deadLetterProducer != nil {
 		if err := consumer.deadLetterProducer.SendDeadLetter(msg); err != nil {
 			log.Error().Err(err).Msg("Failed to load message to dead letter queue")
@@ -208,7 +215,7 @@ func (consumer OCPRulesConsumer) sendDeadLetter(msg *sarama.ConsumerMessage) {
 }
 
 // checkMessageVersion - verifies incoming data's version is expected
-func checkMessageVersion(consumer *OCPRulesConsumer, message *incomingMessage, msg *sarama.ConsumerMessage) {
+func checkMessageVersion(consumer *KafkaConsumer, message *incomingMessage, msg *sarama.ConsumerMessage) {
 	if _, ok := currentSchemaVersion[message.Version]; !ok {
 		warning := fmt.Sprintf("Received data with unexpected version %d.", message.Version)
 		logMessageWarning(consumer, msg, message, warning)
@@ -216,7 +223,7 @@ func checkMessageVersion(consumer *OCPRulesConsumer, message *incomingMessage, m
 }
 
 // checkMessageOrgInAllowList - checks up incoming data's OrganizationID against allowed orgs list
-func checkMessageOrgInAllowList(consumer *OCPRulesConsumer, message *incomingMessage, msg *sarama.ConsumerMessage) (bool, string) {
+func checkMessageOrgInAllowList(consumer *KafkaConsumer, message *incomingMessage, msg *sarama.ConsumerMessage) (bool, string) {
 	if consumer.Configuration.OrgAllowlistEnabled {
 		logMessageInfo(consumer, msg, message, "Checking organization ID against allow list")
 
@@ -232,7 +239,7 @@ func checkMessageOrgInAllowList(consumer *OCPRulesConsumer, message *incomingMes
 	return true, ""
 }
 
-func (consumer *OCPRulesConsumer) writeRecommendations(
+func (consumer *KafkaConsumer) writeRecommendations(
 	msg *sarama.ConsumerMessage, message incomingMessage, reportAsBytes []byte,
 ) (time.Time, error) {
 	err := consumer.Storage.WriteRecommendationsForCluster(
@@ -250,7 +257,7 @@ func (consumer *OCPRulesConsumer) writeRecommendations(
 	return tStored, nil
 }
 
-func (consumer *OCPRulesConsumer) writeInfoReport(
+func (consumer *KafkaConsumer) writeInfoReport(
 	msg *sarama.ConsumerMessage, message incomingMessage, infoStoredAtTime time.Time,
 ) error {
 	// it is expected that message.ParsedInfo contains at least one item:
@@ -273,13 +280,13 @@ func (consumer *OCPRulesConsumer) writeInfoReport(
 	return nil
 }
 
-func (consumer *OCPRulesConsumer) logMsgForFurtherAnalysis(msg *sarama.ConsumerMessage) {
+func (consumer *KafkaConsumer) logMsgForFurtherAnalysis(msg *sarama.ConsumerMessage) {
 	if consumer.Configuration.DisplayMessageWithWrongStructure {
 		log.Info().Str("unparsed message", string(msg.Value)).Msg("Message for further analysis")
 	}
 }
 
-func (consumer *OCPRulesConsumer) logReportStructureError(err error, msg *sarama.ConsumerMessage) {
+func (consumer *KafkaConsumer) logReportStructureError(err error, msg *sarama.ConsumerMessage) {
 	if consumer.Configuration.DisplayMessageWithWrongStructure {
 		log.Err(err).Str("unparsed message", string(msg.Value)).Msg(improperIncomeMessageError)
 	} else {
@@ -287,16 +294,7 @@ func (consumer *OCPRulesConsumer) logReportStructureError(err error, msg *sarama
 	}
 }
 
-func (consumer *OCPRulesConsumer) shouldProcess(consumed *sarama.ConsumerMessage, parsed *incomingMessage) error {
-	err := checkReportStructure(*parsed.Report)
-	if err != nil {
-		consumer.logReportStructureError(err, consumed)
-		return err
-	}
-	return nil
-}
-
-func (consumer *OCPRulesConsumer) retrieveLastCheckedTime(msg *sarama.ConsumerMessage, parsedMsg *incomingMessage) (time.Time, error) {
+func (consumer *KafkaConsumer) retrieveLastCheckedTime(msg *sarama.ConsumerMessage, parsedMsg *incomingMessage) (time.Time, error) {
 	lastCheckedTime, err := time.Parse(time.RFC3339Nano, parsedMsg.LastChecked)
 	if err != nil {
 		logMessageError(consumer, msg, parsedMsg, "Error parsing date from message", err)
@@ -314,101 +312,8 @@ func (consumer *OCPRulesConsumer) retrieveLastCheckedTime(msg *sarama.ConsumerMe
 	return lastCheckedTime, nil
 }
 
-// processMessage processes an incoming message
-func (consumer *OCPRulesConsumer) processMessage(msg *sarama.ConsumerMessage) (types.RequestID, incomingMessage, error) {
-	tStart := time.Now()
-
-	log.Info().Int(offsetKey, int(msg.Offset)).Str(topicKey, consumer.Configuration.Topic).Str(groupKey, consumer.Configuration.Group).Msg("Consumed")
-
-	message, err := consumer.parseMessage(msg, tStart)
-	if err != nil {
-		if err == types.ErrEmptyReport {
-			logMessageInfo(consumer, msg, &message, "This message has an empty report and will not be processed further")
-			metrics.SkippedEmptyReports.Inc()
-			return message.RequestID, message, nil
-		}
-		return message.RequestID, message, err
-	}
-	logMessageInfo(consumer, msg, &message, "Read")
-	tRead := time.Now()
-
-	checkMessageVersion(consumer, &message, msg)
-
-	if ok, cause := checkMessageOrgInAllowList(consumer, &message, msg); !ok {
-		err := errors.New(cause)
-		logMessageError(consumer, msg, &message, cause, err)
-		return message.RequestID, message, err
-	}
-
-	tAllowlisted := time.Now()
-
-	logMessageDebug(consumer, msg, &message, "Marshalled")
-	tMarshalled := time.Now()
-
-	lastCheckedTime, err := consumer.retrieveLastCheckedTime(msg, &message)
-	if err != nil {
-		return message.RequestID, message, err
-	}
-	tTimeCheck := time.Now()
-
-	// timestamp when the report is about to be written into database
-	storedAtTime := time.Now()
-
-	reportAsBytes, err := json.Marshal(*message.Report)
-	if err != nil {
-		logMessageError(consumer, msg, &message, "Error marshalling report", err)
-		return message.RequestID, message, err
-	}
-
-	err = consumer.Storage.WriteReportForCluster(
-		*message.Organization,
-		*message.ClusterName,
-		types.ClusterReport(reportAsBytes),
-		message.ParsedHits,
-		lastCheckedTime,
-		message.Metadata.GatheredAt,
-		storedAtTime,
-		message.RequestID,
-	)
-	if err == types.ErrOldReport {
-		logMessageInfo(consumer, msg, &message, "Skipping because a more recent report already exists for this cluster")
-		return message.RequestID, message, nil
-	} else if err != nil {
-		logMessageError(consumer, msg, &message, "Error writing report to database", err)
-		return message.RequestID, message, err
-	}
-	logMessageDebug(consumer, msg, &message, "Stored report")
-	tStored := time.Now()
-
-	tRecommendationsStored, err := consumer.writeRecommendations(msg, message, reportAsBytes)
-	if err != nil {
-		return message.RequestID, message, err
-	}
-
-	// rule hits has been stored into database - time to log all these great info
-	logClusterInfo(&message)
-
-	infoStoredAtTime := time.Now()
-	if err := consumer.writeInfoReport(msg, message, infoStoredAtTime); err != nil {
-		return message.RequestID, message, err
-	}
-	infoStored := time.Now()
-
-	// log durations for every message consumption steps
-	logDuration(tStart, tRead, msg.Offset, "read")
-	logDuration(tRead, tAllowlisted, msg.Offset, "org_filtering")
-	logDuration(tAllowlisted, tMarshalled, msg.Offset, "marshalling")
-	logDuration(tMarshalled, tTimeCheck, msg.Offset, "time_check")
-	logDuration(tTimeCheck, tStored, msg.Offset, "db_store_report")
-	logDuration(tStored, tRecommendationsStored, msg.Offset, "db_store_recommendations")
-	logDuration(infoStoredAtTime, infoStored, msg.Offset, "db_store_info_report")
-
-	// message has been parsed and stored into storage
-	return message.RequestID, message, nil
-}
-
 // organizationAllowed checks whether the given organization is on allow list or not
-func organizationAllowed(consumer *OCPRulesConsumer, orgID types.OrgID) bool {
+func organizationAllowed(consumer *KafkaConsumer, orgID types.OrgID) bool {
 	allowList := consumer.Configuration.OrgAllowlist
 	if allowList == nil {
 		return false
@@ -417,148 +322,4 @@ func organizationAllowed(consumer *OCPRulesConsumer, orgID types.OrgID) bool {
 	orgAllowed := allowList.Contains(orgID)
 
 	return orgAllowed
-}
-
-func verifySystemAttributeIsEmpty(r Report) bool {
-	var s system
-	if err := json.Unmarshal(*r[reportAttributeSystem], &s); err != nil {
-		return false
-	}
-	if s.Hostname != "" {
-		return false
-	}
-	return true
-}
-
-// isReportWithEmptyAttributes checks if the report is empty, or if the attributes
-// expected in the report, minus the analysis_metadata, are empty.
-// If this function returns true, this report will not be processed further as it is
-// PROBABLY the result of an archive that was not processed by insights-core.
-// see https://github.com/RedHatInsights/insights-results-aggregator/issues/1834
-func isReportWithEmptyAttributes(r Report) bool {
-	// Create attribute checkers for each attribute
-	for attr, attrData := range r {
-		// we don't care about the analysis_metadata attribute
-		if attr == reportAttributeMetadata {
-			continue
-		}
-		// special handling for the system attribute, as it comes with data when empty
-		if attr == reportAttributeSystem {
-			if !verifySystemAttributeIsEmpty(r) {
-				return false
-			}
-			continue
-		}
-		// Check if this attribute of the report is empty
-		checker := JSONAttributeChecker{data: *attrData}
-		if !checker.IsEmpty() {
-			return false
-		}
-	}
-	return true
-}
-
-// checkReportStructure tests if the report has correct structure
-func checkReportStructure(r Report) error {
-	// the structure is not well-defined yet, so all we should do is to check if all keys are there
-
-	// 'skips' key is now optional, we should not expect it anymore:
-	// https://github.com/RedHatInsights/insights-results-aggregator/issues/1206
-	keysNotFound := make([]string, 0, numberOfExpectedKeysInReport)
-	keysFound := 0
-	// check if the structure contains all expected keys
-	for _, expectedKey := range expectedKeysInReport {
-		_, found := r[expectedKey]
-		if !found {
-			keysNotFound = append(keysNotFound, expectedKey)
-		} else {
-			keysFound++
-		}
-	}
-
-	if keysFound == numberOfExpectedKeysInReport {
-		return nil
-	}
-
-	// empty reports mean that this message should not be processed further
-	isEmpty := len(r) == 0 || isReportWithEmptyAttributes(r)
-	if isEmpty {
-		log.Debug().Msg("Empty report or report with only empty attributes. Processing of this message will be skipped.")
-		return types.ErrEmptyReport
-	}
-
-	// report is not empty, and some keys have not been found -> malformed
-	if len(keysNotFound) != 0 {
-		return fmt.Errorf("improper report structure, missing key(s) with name '%v'", keysNotFound)
-	}
-
-	return nil
-}
-
-// deserializeMessage tries to parse incoming message and read all required attributes from it
-func deserializeMessage(messageValue []byte) (incomingMessage, error) {
-	var deserialized incomingMessage
-
-	err := json.Unmarshal(messageValue, &deserialized)
-	if err != nil {
-		return deserialized, err
-	}
-
-	if deserialized.Organization == nil {
-		return deserialized, errors.New("missing required attribute 'OrgID'")
-	}
-	if deserialized.ClusterName == nil {
-		return deserialized, errors.New("missing required attribute 'ClusterName'")
-	}
-	if deserialized.Report == nil {
-		return deserialized, errors.New("missing required attribute 'Report'")
-	}
-
-	_, err = uuid.Parse(string(*deserialized.ClusterName))
-
-	if err != nil {
-		return deserialized, errors.New("cluster name is not a UUID")
-	}
-	return deserialized, nil
-}
-
-// parseReportContent verifies the content of the Report structure and parses it into
-// the relevant parts of the incomingMessage structure
-func parseReportContent(message *incomingMessage) error {
-	err := json.Unmarshal(*((*message.Report)[reportAttributeReports]), &message.ParsedHits)
-	if err != nil {
-		return err
-	}
-
-	// it is expected that message.ParsedInfo contains at least one item:
-	// result from special INFO rule containing cluster version that is
-	// used just in external data pipeline
-	err = json.Unmarshal(*((*message.Report)[reportAttributeInfo]), &message.ParsedInfo)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (consumer *OCPRulesConsumer) parseMessage(msg *sarama.ConsumerMessage, tStart time.Time) (incomingMessage, error) {
-	message, err := deserializeMessage(msg.Value)
-	if err != nil {
-		consumer.logMsgForFurtherAnalysis(msg)
-		logUnparsedMessageError(consumer, msg, "Error parsing message from Kafka", err)
-		return message, err
-	}
-
-	consumer.updatePayloadTracker(message.RequestID, tStart, message.Organization, message.Account, producer.StatusReceived)
-
-	if err := consumer.shouldProcess(msg, &message); err != nil {
-		return message, err
-	}
-
-	err = parseReportContent(&message)
-	if err != nil {
-		consumer.logReportStructureError(err, msg)
-		return message, err
-	}
-
-	return message, nil
 }
